refactor(btcio): create corepc_node_helpers
