+++ crates/btcio/src/test_utils.rs
@@ -236,3 +236,37 @@ pub fn build_reveal_transaction_test(
         control_block,
     )
 }
+
+#[cfg(test)]
+pub mod corepc_node_helpers {
+    use bitcoin::{Address, BlockHash};
+    use corepc_node::BitcoinD;
+
+    /// Get the authentication credentials for a given `bitcoind` instance.
+    pub fn get_auth(bitcoind: &BitcoinD) -> (String, String) {
+        let params = &bitcoind.params;
+        let cookie_values = params.get_cookie_values().unwrap().unwrap();
+        (cookie_values.user, cookie_values.password)
+    }
+
+    /// Mine a number of blocks of a given size `count`, which may be specified to a given coinbase
+    /// `address`.
+    pub fn mine_blocks(
+        bitcoind: &BitcoinD,
+        count: usize,
+        address: Option<Address>,
+    ) -> anyhow::Result<Vec<BlockHash>> {
+        let coinbase_address = match address {
+            Some(address) => address,
+            None => bitcoind.client.new_address()?,
+        };
+        let block_hashes = bitcoind
+            .client
+            .generate_to_address(count as _, &coinbase_address)?
+            .0
+            .iter()
+            .map(|hash| hash.parse::<BlockHash>())
+            .collect::<Result<Vec<_>, _>>()?;
+        Ok(block_hashes)
+    }
+}
